mobile(sdk): split HonuaMobileClient into per-feature partials + cite #199 in orphan skip
Split the 1504-line HonuaMobileClient.cs into per-feature partial files to
address the god-class concern raised in fresh-eyes review. No public API or
behavior changes — verified via ilspycmd public-surface diff.

Partial files (line counts):
- HonuaMobileClient.cs            240 — ctors, properties, Dispose, options builders, message handler
- HonuaMobileClient.Features.cs   557 — Query/ApplyEdits (REST + SDK contracts), edit-response converters
- HonuaMobileClient.Internals.cs  296 — SendJsonAsync, URI/auth helpers, JSON Try* helpers
- HonuaMobileClient.Attachments.cs 230 — list/download/add/update/delete + REST-compat fallback
- HonuaMobileClient.OgcFeatures.cs 145 — collections + item CRUD via OGC contract
- HonuaMobileClient.Grpc.cs        81 — lazy gRPC client, options/provider builders

Also updates the lone orphan Skip in HonuaMobileSdkFeatureClientTests to cite
issue #199 (same root cause as the existing live-server SkippableFact gate)
so every non-env-gated skip points at a tracking issue.
