+++ src/Honua.Mobile.Sdk/HonuaMobileClient.OgcFeatures.cs
@@ -1,0 +1,145 @@
+using System.Text.Json;
+using Honua.Sdk.Abstractions.Features;
+using Honua.Sdk.OgcFeatures.Exceptions;
+
+using OgcRequestConverters = Honua.Sdk.OgcFeatures.Conversion.RequestConverters;
+
+namespace Honua.Mobile.Sdk;
+
+public sealed partial class HonuaMobileClient
+{
+    /// <summary>
+    /// Retrieves the list of OGC Features API collections available on the server.
+    /// </summary>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>A <see cref="JsonDocument"/> describing available collections.</returns>
+    /// <exception cref="HonuaMobileApiException">Thrown when the server returns a non-success HTTP status code.</exception>
+    public async Task<JsonDocument> GetOgcCollectionsAsync(CancellationToken ct = default)
+    {
+        try
+        {
+            var collections = await _ogcFeaturesClient.ListCollectionsAsync(ct).ConfigureAwait(false);
+            return OgcRequestConverters.ToJsonDocument(collections);
+        }
+        catch (HonuaOgcFeaturesException ex)
+        {
+            throw ToMobileApiException("OGC Features", ex);
+        }
+    }
+
+    /// <summary>
+    /// Retrieves items from an OGC Features API collection with optional filtering and pagination.
+    /// </summary>
+    /// <param name="request">Parameters including collection ID, CQL filter, limit, and offset.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>A <see cref="JsonDocument"/> containing the matched items.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when <paramref name="request"/> is <see langword="null"/>.</exception>
+    /// <exception cref="HonuaMobileApiException">Thrown when the server returns a non-success HTTP status code.</exception>
+    public async Task<JsonDocument> GetOgcItemsAsync(OgcItemsRequest request, CancellationToken ct = default)
+    {
+        ArgumentNullException.ThrowIfNull(request);
+
+        try
+        {
+            var response = await _ogcFeaturesClient
+                .GetItemsAsync(request.CollectionId, OgcRequestConverters.ToOgcItemsParams(request), ct)
+                .ConfigureAwait(false);
+            return OgcRequestConverters.ToJsonDocument(response);
+        }
+        catch (HonuaOgcFeaturesException ex)
+        {
+            throw ToMobileApiException("OGC Features", ex);
+        }
+    }
+
+    /// <summary>
+    /// Creates a new feature item in an OGC Features API collection.
+    /// </summary>
+    /// <param name="request">The collection ID and GeoJSON feature to create.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>A <see cref="JsonDocument"/> containing the server response for the created item.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when <paramref name="request"/> is <see langword="null"/>.</exception>
+    /// <exception cref="HonuaMobileApiException">Thrown when the server returns a non-success HTTP status code.</exception>
+    public async Task<JsonDocument> CreateOgcItemAsync(OgcCreateItemRequest request, CancellationToken ct = default)
+    {
+        ArgumentNullException.ThrowIfNull(request);
+
+        try
+        {
+            var response = await _ogcFeaturesClient
+                .CreateItemAsync(request.CollectionId, OgcRequestConverters.ToOgcFeature(request.Feature), ct)
+                .ConfigureAwait(false);
+            return OgcRequestConverters.ToJsonDocument(response);
+        }
+        catch (HonuaOgcFeaturesException ex)
+        {
+            throw ToMobileApiException("OGC Features", ex);
+        }
+    }
+
+    /// <summary>
+    /// Replaces an existing feature item in an OGC Features API collection (HTTP PUT).
+    /// </summary>
+    /// <param name="request">The collection ID, feature ID, and full replacement feature.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>A <see cref="JsonDocument"/> containing the server response.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when <paramref name="request"/> is <see langword="null"/>.</exception>
+    /// <exception cref="HonuaMobileApiException">Thrown when the server returns a non-success HTTP status code.</exception>
+    public async Task<JsonDocument> ReplaceOgcItemAsync(OgcReplaceItemRequest request, CancellationToken ct = default)
+    {
+        ArgumentNullException.ThrowIfNull(request);
+
+        try
+        {
+            var response = await _ogcFeaturesClient
+                .UpdateItemAsync(request.CollectionId, request.FeatureId, OgcRequestConverters.ToOgcFeature(request.Feature), ct)
+                .ConfigureAwait(false);
+            return OgcRequestConverters.ToJsonDocument(response);
+        }
+        catch (HonuaOgcFeaturesException ex)
+        {
+            throw ToMobileApiException("OGC Features", ex);
+        }
+    }
+
+    /// <summary>
+    /// Partially updates an existing feature item using JSON Merge Patch (RFC 7396).
+    /// </summary>
+    /// <param name="request">The collection ID, feature ID, and merge-patch document.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>A <see cref="JsonDocument"/> containing the server response.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when <paramref name="request"/> is <see langword="null"/>.</exception>
+    /// <exception cref="HonuaMobileApiException">Thrown when the server returns a non-success HTTP status code.</exception>
+    public async Task<JsonDocument> PatchOgcItemAsync(OgcPatchItemRequest request, CancellationToken ct = default)
+    {
+        ArgumentNullException.ThrowIfNull(request);
+
+        try
+        {
+            var response = await _ogcFeaturesClient
+                .PatchItemAsync(request.CollectionId, request.FeatureId, OgcRequestConverters.ToJsonElement(request.Patch), ct)
+                .ConfigureAwait(false);
+            return OgcRequestConverters.ToJsonDocument(response);
+        }
+        catch (HonuaOgcFeaturesException ex)
+        {
+            throw ToMobileApiException("OGC Features", ex);
+        }
+    }
+
+    /// <summary>
+    /// Deletes a feature item from an OGC Features API collection.
+    /// </summary>
+    /// <param name="request">The collection ID and feature ID to delete.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>A <see cref="JsonDocument"/> containing the server response.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when <paramref name="request"/> is <see langword="null"/>.</exception>
+    /// <exception cref="HonuaMobileApiException">Thrown when the server returns a non-success HTTP status code.</exception>
+    public async Task<JsonDocument> DeleteOgcItemAsync(OgcDeleteItemRequest request, CancellationToken ct = default)
+    {
+        ArgumentNullException.ThrowIfNull(request);
+
+        var path = $"/ogc/features/collections/{Uri.EscapeDataString(request.CollectionId)}/items/{Uri.EscapeDataString(request.FeatureId)}";
+        return await SendJsonAsync(HttpMethod.Delete, path, null, null, ct).ConfigureAwait(false);
+    }
+}
